requirements.txt
requests
pandas
playwright
supabase
python-dotenv
sentence-transformers